requirements.txt
# FastAPI 및 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 데이터베이스
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# WebSocket
websockets==12.0
aiohttp==3.9.1
httpx==0.25.2

# 유틸리티
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# 통계 계산
numpy==1.26.2
pandas==2.1.4

# 로깅
structlog==23.2.0

# 테스트
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# 타입 체크
mypy==1.7.1
types-python-dateutil==2.8.19.14

# MCP 서버
fastmcp


